+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.15.1
 
+- Fix: Incorrect axes scale domain ([#107](https://github.com/flekschas/jupyter-scatter/issues/107))
 - Fix: Use custom regl-scatterplot option on creating a `Scatter` instance ([#106](https://github.com/flekschas/jupyter-scatter/issues/106))
 
 ## v0.15.0